Fix VS2015 build warning

+++ src/Core/Registration/Feature.cpp
@@ -73,7 +73,7 @@ std::shared_ptr<Feature> ComputeSPFHFeature(const PointCloud &input,
 		const KDTreeFlann &kdtree, const KDTreeSearchParam &search_param)
 {
 	auto feature = std::make_shared<Feature>();
-	feature->Resize(33, input.points_.size());
+	feature->Resize(33, (int)input.points_.size());
 #ifdef _OPENMP
 #pragma omp parallel for schedule(static)
 #endif
@@ -113,7 +113,7 @@ std::shared_ptr<Feature> ComputeFPFHFeature(const PointCloud &input,
 		const KDTreeSearchParam &search_param/* = KDTreeSearchParamKNN()*/)
 {
 	auto feature = std::make_shared<Feature>();
-	feature->Resize(33, input.points_.size());
+	feature->Resize(33, (int)input.points_.size());
 	if (input.HasNormals() == false) {
 		PrintDebug("[ComputeFPFHFeature] Failed because input point cloud has no normal.\n");
 		return feature;
